Modify: add tutorialData in the same format with api response
